fix(symbolic): IfcMappedItem resolution failures must not read as Y=0
STACKED ON #2352 -- needs its `Transform2D::unresolved()`. Review/merge
that one first.

#2352 established the shape: in `symbolic/transform.rs`, returning
`Transform2D::identity()` on a GENUINE FAILURE was indistinguishable from
a legitimate ground-floor elevation, because `identity().tz == 0.0`.
`items.rs`'s `IfcMappedItem` branch has the identical shape and was left
out of #2352 only because the file sat exactly AT its module-size ratchet
ceiling (440/440). This closes it.

FOUR CALL SITES, ALL MANDATORY ATTRIBUTES
-----------------------------------------
- `rep_map.get_ref(0)` = `IfcRepresentationMap.MappingOrigin`
- `item.get_ref(1)`    = `IfcMappedItem.MappingTarget`

Both the `Err(_)` (dangling ref) and `None` (absent/null) arms of each now
return `unresolved()`. Attribute INDICES verified against the generated
schema rather than assumed -- `rust/core/src/generated/schema.rs:11510`
lists `IfcRepresentationMap => ["MappingOrigin", "MappedRepresentation"]`
(index 0) and `:9843` lists `IfcMappedItem => ["MappingSource",
"MappingTarget"]` (index 1).

Both attributes are MANDATORY in the schema, so absence is malformed data
rather than a legitimate default. That is the distinction #2352 drew: it
deliberately KEPT `identity()` for `ObjectPlacement`, which IS optional on
`IfcProduct` (`transform.rs:114-119`, absent/null -> identity). Verified
that treatment is still correct rather than copying the new pattern
blindly.

Deliberately NOT changed: the dangling `MappingSource` path at the top of
the branch. It early-returns and emits nothing at all, so there is no
elevation to be ambiguous about -- a "drop the item" failure, not a
"silently poison to 0.0" one.

RATCHET: SPLIT, NOT RAISED
--------------------------
The fix alone pushed `items.rs` to 444 against its 440 ceiling. No budget
was raised. `extract_trimmed_curve` (a self-contained `IfcTrimmedCurve`
tessellation helper with no other callers) moved to a new sibling module
`symbolic/trimmed_curve.rs`, following this directory's existing
per-concern-file convention (`fill.rs`, `grid.rs`, `text.rs`).

`items.rs` is now 323 lines -- under the global LIMIT of 400 -- so its
allowlist row was DELETED entirely rather than adjusted, per the ratchet's
own documented rule.

The move was verified to be a PURE MOVE by extracting the function from
both the before and after files and comparing: 117 lines, byte-identical
modulo the visibility keyword that a cross-module move requires. The
split therefore carries no behaviour risk of its own.

RED PROOF
---------
With the four `unresolved()` calls reverted to `identity()` (exact
substring replacement, count asserted == 4):

  test genuine_zero_elevation_mapped_item_stays_finite_zero ... ok
  test dangling_mapping_origin_is_not_finite ... FAILED
  test dangling_mapping_target_is_not_finite ... FAILED
  ... world_y must be non-finite (NaN), not silently 0.0 ... Got: 0
  test result: FAILED. 1 passed; 2 failed

Restored by file copy and proven byte-identical with `diff`.

BOUNDING CONTROL (passes before AND after)
------------------------------------------
`genuine_zero_elevation_mapped_item_stays_finite_zero` -- a mapped item
whose MappingOrigin and MappingTarget both resolve cleanly to Z=0 still
yields exactly `world_y == 0.0`. Note it passes in the RED run above, as a
bounding control must. Without it, "return NaN unconditionally" would
satisfy both failure tests.

DISPLACEMENT
------------
This widens NaN to a rarer, deeper case on the SAME field #2352 already
swept (`SymbolicPolyline.world_y` and its circle/text equivalents), through
the identical pipe, so no new consumer is exposed. Re-confirmed the two
that matter: `symbolic-parse.ts:119` gates on `Number.isFinite` and is
source-agnostic, and the renderer never receives a raw primitive Y --
`useSymbolicAnnotations` passes `resolveBucketY(bucket.storeyElevation,
fallbackY)`, both finite. That second one is the important one: the fill
upload feeds `expandModelBoundsWithFlatVertices` -> `camera.setSceneBounds`,
where a NaN would poison model bounds for the whole scene.

VERIFICATION
------------
`rust/processing` full `cargo test`: 47 test binaries, all green, exit 0,
module-size ratchet included (scratch CARGO_TARGET_DIR throughout).
+3 tests. No changeset -- Rust crate, not an npm package.

THE SHAPE IS NOT YET FULLY CLOSED -- enumerated, not estimated
--------------------------------------------------------------
Six `Transform2D::identity()`-on-failure sites remain in this directory:
`text.rs:38`, `text.rs:40`, and `mod.rs:202,204,209,211`. They are NOT
fixed here and are being handled separately rather than silently widening
this diff. The three remaining sites in `transform.rs` (115, 118, 145) are
CORRECT as-is: those are the optional-`ObjectPlacement` and
top-of-placement-chain cases.
